[fix] 지출 통계 내역퍼센트 수정

+++ src/main/java/com/sunny/backend/dto/response/consumption/SpendTypeStatisticsResponse.java
@@ -13,6 +13,6 @@ public class SpendTypeStatisticsResponse {
     private SpendType category;
     private long totalCount;
     private long totalMoney;
-    private long percentage;
+    private double percentage;
 
 }
